% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-to-html-table.R
\name{dfToHtmlTable}
\alias{dfToHtmlTable}
\title{Dataframe to HTML Table}
\usage{
dfToHtmlTable(
  df,
  align = "c",
  width = "100\%",
  font = "\\"Sintony\\", Arial, sans-serif",
  headerBgColour = "#15679f",
  headerFontColour = "#ffffff",
  extraHeaderCss = NA,
  strippedBgColour = "#cccccc",
  strippedFontColour = "#000000",
  extraRowCss = NA,
  highlightRowColour = NULL,
  highlightRows = NULL,
  borderStyle = "1px solid black",
  borderLocation = "all",
  colToCollapse = NULL
)
}
\arguments{
\item{df}{A data.frame}
}
\value{

}
\description{
Simple HTML Table intended to be emailed.
}
\examples{
htmltools::html_print(dfToHtmlTable(head(mtcars)))
htmltools::html_print(dfToHtmlTable(iris[0, ]))
htmltools::html_print(dfToHtmlTable(iris, colToCollapse = "Species"))
htmltools::html_print(dfToHtmlTable(mtcars, colToCollapse = "carb"))
}
