% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-notin-inflex.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In operator}
\usage{
x \%notin\% table
}
\description{
Not In operator
}
\examples{
"a" \%notin\% LETTERS
"A" \%notin\% LETTERS
1 \%notin\% NULL
}
\keyword{internal}
