% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paste-replace-na.R
\name{pasteReplaceNa}
\alias{pasteReplaceNa}
\title{Paste Replace NA}
\usage{
pasteReplaceNa(...)
}
\arguments{
\item{...}{args to turn into a string}
}
\value{

}
\description{
Similar to base r paste0 but remove args that are NA
}
\examples{
pasteReplaceNa("a", NA, 1)
pasteReplaceNa("a", "b")
}
