% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{map}
\alias{keep}
\alias{pluck}
\alias{tags}
\alias{str_glue}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmltools}{\code{\link[htmltools:builder]{tags}}}

  \item{purrr}{\code{\link[purrr]{keep}}, \code{\link[purrr]{map}}, \code{\link[purrr]{pluck}}}

  \item{stringr}{\code{\link[stringr]{str_glue}}}
}}

