% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-to-html-table.R
\name{dfToHtmlTable}
\alias{dfToHtmlTable}
\title{Dataframe to HTML Table}
\usage{
dfToHtmlTable(
  df,
  align = "c",
  width = "100\%",
  font = "Arial, sans-serif",
  headerBgColour = "#15679f",
  headerFontColour = "#ffffff",
  headerFontSize = "12px",
  headerFontWeight = "bold",
  strippedBgColour = "#cccccc",
  strippedFontColour = "#000000",
  cellFontSize = "12px",
  highlightRowColour = NULL,
  highlightRows = NULL,
  borderStyle = "1px solid black",
  borderLocation = "all",
  colToCollapse = NULL,
  extraHeaderCss = NA,
  extraRowCss = NA,
  nowrapCols = NULL,
  title = NULL,
  subtitle = NULL,
  footnotes = NULL,
  titleAlignment = "left"
)
}
\arguments{
\item{df}{A data.frame}

\item{align}{string to determine alignment eg. "lcr". last character to used to fill remainder}

\item{width}{table width}

\item{font}{font family}

\item{headerBgColour}{header bg colour}

\item{headerFontColour}{header font colour}

\item{headerFontSize}{header font size}

\item{headerFontWeight}{header font weight}

\item{strippedBgColour}{stripped row bg colour}

\item{strippedFontColour}{stripped row font colour}

\item{cellFontSize}{cell font size}

\item{highlightRowColour}{highlight row colour}

\item{highlightRows}{highlight row indices (positions)}

\item{borderStyle}{css for border style, eg. "1px solid black"}

\item{borderLocation}{border location, can be "all", "row", or "col"}

\item{colToCollapse}{col to collapse on. this applies row spans and orders by the column
so that the value does not repeat}

\item{extraHeaderCss}{extra header css}

\item{extraRowCss}{extra row css}

\item{nowrapCols}{cols to apply nowrap to. see css white-space nowrap.}

\item{title}{table title}

\item{subtitle}{table subtitle, requires title}

\item{footnotes}{table foot notes}

\item{titleAlignment}{title alignment}
}
\value{

}
\description{
Simple HTML Table intended to be emailed.
}
\examples{
htmltools::html_print(dfToHtmlTable(head(mtcars)))
htmltools::html_print(dfToHtmlTable(iris[0, ]))
htmltools::html_print(dfToHtmlTable(iris, colToCollapse = "Species"))
htmltools::html_print(dfToHtmlTable(mtcars, colToCollapse = "carb"))
}
