% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-row-collapse.R
\name{rowCollapse}
\alias{rowCollapse}
\title{Collapse rows in HTML table}
\usage{
rowCollapse(htmlTable, colToCollapse, strippedBgColour)
}
\arguments{
\item{htmlTable}{html table made with htmltools::tags}

\item{colToCollapse}{column name to collapse}

\item{strippedBgColour}{collapsed row background colour}
}
\value{

}
\description{
Collapse rows in HTML table
}
